+++ Scripts/Global.gd
@@ -111,10 +111,11 @@ var show_guides := true
 var show_animation_timeline := true
 
 # Onion skinning options
+var onion_skinning := false
 # warning-ignore:unused_class_variable
-var onion_skinning_past_rate := 0
+var onion_skinning_past_rate := 1
 # warning-ignore:unused_class_variable
-var onion_skinning_future_rate := 0
+var onion_skinning_future_rate := 1
 # warning-ignore:unused_class_variable
 var onion_skinning_blue_red := false
 
@@ -188,14 +189,10 @@ var left_brush_type_container : Container
 var right_brush_type_container : Container
 var left_brush_type_button : BaseButton
 var right_brush_type_button : BaseButton
-var left_brush_type_label : Label
-var right_brush_type_label : Label
 var brushes_popup : Popup
 var file_brush_container : GridContainer
 var project_brush_container : GridContainer
 
-var left_brush_size_container : Container
-var right_brush_size_container : Container
 var left_brush_size_edit : SpinBox
 var left_brush_size_slider : HSlider
 var right_brush_size_edit : SpinBox
@@ -307,41 +304,37 @@ func _ready() -> void:
 	right_brush_type_container = find_node_by_name(right_tool_options_container, "RightBrushType")
 	left_brush_type_button = find_node_by_name(left_brush_type_container, "LeftBrushTypeButton")
 	right_brush_type_button = find_node_by_name(right_brush_type_container, "RightBrushTypeButton")
-	left_brush_type_label = find_node_by_name(left_brush_type_container, "LeftBrushTypeLabel")
-	right_brush_type_label = find_node_by_name(right_brush_type_container, "RightBrushTypeLabel")
 	brushes_popup = find_node_by_name(root, "BrushesPopup")
 	file_brush_container = find_node_by_name(brushes_popup, "FileBrushContainer")
 	project_brush_container = find_node_by_name(brushes_popup, "ProjectBrushContainer")
 
-	left_brush_size_container = find_node_by_name(left_tool_options_container, "LeftBrushSize")
-	right_brush_size_container = find_node_by_name(right_tool_options_container, "RightBrushSize")
-	left_brush_size_edit = find_node_by_name(left_brush_size_container, "LeftBrushSizeEdit")
-	left_brush_size_slider = find_node_by_name(left_brush_size_container, "LeftBrushSizeSlider")
-	right_brush_size_edit = find_node_by_name(right_brush_size_container, "RightBrushSizeEdit")
-	right_brush_size_slider = find_node_by_name(right_brush_size_container, "RightBrushSizeSlider")
+	left_brush_size_edit = find_node_by_name(root, "LeftBrushSizeEdit")
+	left_brush_size_slider = find_node_by_name(root, "LeftBrushSizeSlider")
+	right_brush_size_edit = find_node_by_name(root, "RightBrushSizeEdit")
+	right_brush_size_slider = find_node_by_name(root, "RightBrushSizeSlider")
 
-	left_color_interpolation_container = find_node_by_name(left_tool_options_container, "LeftColorInterpolation")
-	right_color_interpolation_container = find_node_by_name(right_tool_options_container, "RightColorInterpolation")
-	left_interpolate_spinbox = find_node_by_name(left_color_interpolation_container, "LeftInterpolateFactor")
-	left_interpolate_slider = find_node_by_name(left_color_interpolation_container, "LeftInterpolateSlider")
-	right_interpolate_spinbox = find_node_by_name(right_color_interpolation_container, "RightInterpolateFactor")
-	right_interpolate_slider = find_node_by_name(right_color_interpolation_container, "RightInterpolateSlider")
+	left_color_interpolation_container = find_node_by_name(root, "LeftColorInterpolation")
+	right_color_interpolation_container = find_node_by_name(root, "RightColorInterpolation")
+	left_interpolate_spinbox = find_node_by_name(root, "LeftInterpolateFactor")
+	left_interpolate_slider = find_node_by_name(root, "LeftInterpolateSlider")
+	right_interpolate_spinbox = find_node_by_name(root, "RightInterpolateFactor")
+	right_interpolate_slider = find_node_by_name(root, "RightInterpolateSlider")
 
-	left_fill_area_container = find_node_by_name(left_tool_options_container, "LeftFillArea")
-	right_fill_area_container = find_node_by_name(right_tool_options_container, "RightFillArea")
+	left_fill_area_container = find_node_by_name(root, "LeftFillArea")
+	right_fill_area_container = find_node_by_name(root, "RightFillArea")
 
-	left_ld_container = find_node_by_name(left_tool_options_container, "LeftLDOptions")
-	left_ld_amount_slider = find_node_by_name(left_ld_container, "LeftLDAmountSlider")
-	left_ld_amount_spinbox = find_node_by_name(left_ld_container, "LeftLDAmountSpinbox")
-	right_ld_container = find_node_by_name(right_tool_options_container, "RightLDOptions")
-	right_ld_amount_slider = find_node_by_name(right_ld_container, "RightLDAmountSlider")
-	right_ld_amount_spinbox = find_node_by_name(right_ld_container, "RightLDAmountSpinbox")
+	left_ld_container = find_node_by_name(root, "LeftLDOptions")
+	left_ld_amount_slider = find_node_by_name(root, "LeftLDAmountSlider")
+	left_ld_amount_spinbox = find_node_by_name(root, "LeftLDAmountSpinbox")
+	right_ld_container = find_node_by_name(root, "RightLDOptions")
+	right_ld_amount_slider = find_node_by_name(root, "RightLDAmountSlider")
+	right_ld_amount_spinbox = find_node_by_name(root, "RightLDAmountSpinbox")
 
-	left_colorpicker_container = find_node_by_name(left_tool_options_container, "LeftColorPickerOptions")
-	right_colorpicker_container = find_node_by_name(right_tool_options_container, "RightColorPickerOptions")
+	left_colorpicker_container = find_node_by_name(root, "LeftColorPickerOptions")
+	right_colorpicker_container = find_node_by_name(root, "RightColorPickerOptions")
 
-	left_mirror_container = find_node_by_name(left_tool_options_container, "LeftMirroring")
-	right_mirror_container = find_node_by_name(right_tool_options_container, "RightMirroring")
+	left_mirror_container = find_node_by_name(root, "LeftMirrorButtons")
+	right_mirror_container = find_node_by_name(root, "RightMirrorButtons")
 
 	animation_timeline = find_node_by_name(root, "AnimationTimeline")
 
@@ -356,7 +349,7 @@ func _ready() -> void:
 
 	remove_layer_button = find_node_by_name(animation_timeline, "RemoveLayer")
 	move_up_layer_button = find_node_by_name(animation_timeline, "MoveUpLayer")
-	move_down_layer_button = find_node_by_name(animation_timeline, "MovwDownLayer")
+	move_down_layer_button = find_node_by_name(animation_timeline, "MoveDownLayer")
 	merge_down_layer_button = find_node_by_name(animation_timeline, "MergeDownLayer")
 
 	layer_opacity_slider = find_node_by_name(animation_timeline, "OpacitySlider")
@@ -373,8 +366,9 @@ func _ready() -> void:
 
 	error_dialog = find_node_by_name(root, "ErrorDialog")
 
-	# Store [Layer name, Layer visibility boolean, Layer lock boolean, Frame container]
-	layers.append([tr("Layer") + " 0", true, false, HBoxContainer.new()])
+	# Store [Layer name (0), Layer visibility boolean (1), Layer lock boolean (2), Frame container (3),
+	# will new frames be linked boolean (4), Array of linked frames (5)]
+	layers.append([tr("Layer") + " 0", true, false, HBoxContainer.new(), false, []])
 
 # Thanks to https://godotengine.org/qa/17524/how-to-find-an-instanced-scene-by-its-name
 func find_node_by_name(root, node_name) -> Node: